fraud_investigator/requirements.txt
streamlit
langchain
langchain-openai
langchain-aws
langchain-google-genai
langchain-community
langgraph
pydantic
python-dotenv
pandas
